Move logic out of SymbolIndexerTest to index.InverseSymbolIndexer
